Resource: changed driver defs, removed init_at
The driver definitions have been changed such that the generic driver
definition features the stat function. This way stats aren't acquired
from super block drivers. If the filename given to a stat function is
NULL than the current resource is statted, if the filename is non-NULL
then it corresponds to a path within the resource, and the stat of that
inner resource will be returned. There is now a control function as
well, which allows for a caller to submit a buffer to the driver to
configure it in some way. The init_resource_at functions have also been
removed as device drivers may now create their own files instead of
having the caller worry about them.