Added event system
The implementation of the system is very similar to the ListMutex
implementation; however, instead of yielding to a thread, a specified
function is called. Though it may be valuable to change these function
pointers for threads such that a specific context can be jumped into, or
to at least give that option.