graph: first check of semantics in base
I have went over the order in which the semaphores are incremented and
decremented, how values are read, and how values are used. I have made a
few corrections: ref_counts are only used to indicate the number of
operations the node is currently being used by. When recurisvely
deleting a node, the ref_count is now atomically loaded use
ARC_ATOMIC_LOAD. This macro uses acquire semantics and should take place
after atomic stores. When traversing the next node's ref_count is
incremented prior to the current node's being decremented -
hand-over-hand locking. Lastly, in traversal an additional check was
added to ensure that the current node is different from the last node.
If it is not, then the traversal has been trapped by a removal
operation. This case needs to be limited by changing how removal works.