graph: base review two
I have added in a ref_count check to the removal function. This should
hopefully catch traversals and prevent them from being split off.
Potentially a measure could be added to forcefully remove nodes, but I
think this can be acheived through other means by the caller. When
adding a node, I forgot to set its parent and name. However, I have an
idea for how the naming can be done in a nice way.