graph: base names
The name attribute is now set. This is done on the add function to make
it easier to move around and rename nodes. The logic behind naming is:
* If no _name is specified use the empty name if a name has not already
been specified
* If a _name is give, use that name always
I have also added in the freeing logic. If the name is an empty name, it
will not be freed. The create function has also been exposed.