path: flip to/from, fix path_get_*
I have changed the order of to and from, doing (to, from) got a little
annoying. Further, I have fixed the path_get_* functions - it was really
only path_get_abs. It now inserts the separator character after the
component name, and sets the last separator to 0. I also forgot to
switch to set current to parent when traversing up the node graph.